Cleaned up the old highlighting mechanism
