+++ src/main/java/org/antlr/jetbrains/st4plugin/highlight/STSyntaxHighlighter.java
@@ -1,102 +1,59 @@
 package org.antlr.jetbrains.st4plugin.highlight;
 
+import com.intellij.lexer.Lexer;
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors;
-import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.colors.TextAttributesKey;
-import org.antlr.jetbrains.st4plugin.parsing.ParserErrorListener;
-import org.antlr.jetbrains.st4plugin.parsing.ParsingResult;
+import com.intellij.openapi.fileTypes.SyntaxHighlighterBase;
+import com.intellij.psi.tree.IElementType;
+import org.antlr.intellij.adaptor.lexer.ANTLRLexerAdaptor;
+import org.antlr.jetbrains.st4plugin.STLanguage;
 import org.antlr.jetbrains.st4plugin.parsing.STLexer;
-import org.antlr.jetbrains.st4plugin.parsing.STParser;
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.Lexer;
-import org.antlr.v4.runtime.Parser;
-import org.antlr.v4.runtime.ParserRuleContext;
-import org.antlr.v4.runtime.Token;
-import org.antlr.v4.runtime.tree.ParseTree;
-import org.antlr.v4.runtime.tree.TerminalNode;
-import org.antlr.v4.runtime.tree.xpath.XPath;
+import org.antlr.jetbrains.st4plugin.psi.STTokenTypes;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static com.intellij.openapi.editor.colors.TextAttributesKey.createTextAttributesKey;
 
-public class STSyntaxHighlighter extends SyntaxHighlighter {
-	public static final TextAttributesKey STGroup_TEMPLATE_TEXT =
-		createTextAttributesKey("STGroup_TEMPLATE_TEXT", DefaultLanguageHighlighterColors.TEMPLATE_LANGUAGE_COLOR);
-	public static final TextAttributesKey ST_ID = STGroupSyntaxHighlighter.STGroup_TEMPLATE_NAME;
+public class STSyntaxHighlighter extends SyntaxHighlighterBase {
 
-	public STGroupSyntaxHighlighter groupHighlighter;
-	public Token templateToken;
+    public static final TextAttributesKey STGroup_TEMPLATE_TEXT =
+            createTextAttributesKey("STGroup_TEMPLATE_TEXT", DefaultLanguageHighlighterColors.TEMPLATE_LANGUAGE_COLOR);
+    public static final TextAttributesKey ST_ID = STGroupSyntaxHighlighter.STGroup_TEMPLATE_NAME;
 
-	public STSyntaxHighlighter(STGroupSyntaxHighlighter groupHighlighter,
-	                           Editor editor,
-	                           Token templateToken,
-	                           int startIndex)
-	{
-		super(editor, startIndex, groupHighlighter.highlightInfos);
-		this.templateToken = templateToken;
-		this.groupHighlighter = groupHighlighter;
-	}
+    private static final List<IElementType> KEYWORDS = Stream.of(
+            STLexer.IF, STLexer.ELSE, /* TODO STLexer.REGION_END, */ STLexer.TRUE,
+            STLexer.FALSE, STLexer.ELSEIF, STLexer.ENDIF, STLexer.SUPER
+    ).map(STTokenTypes::getTokenElementType).collect(Collectors.toList());
 
-	@Override
-	public Lexer getLexer(String text) {
-//		STLexer lexer = new STLexer(new ANTLR3StringStream(text), templateToken, groupHighlighter.delimiters[0], groupHighlighter.delimiters[1]);
-//		return lexer;
-		return null;
-	}
+    @NotNull
+    @Override
+    public Lexer getHighlightingLexer() {
+        return new ANTLRLexerAdaptor(STLanguage.INSTANCE, new STLexer(null));
+    }
 
-	@Override
-	public ParsingResult parse(CommonTokenStream tokens) {
-		STParser parser = new STParser(tokens);
-		parser.removeErrorListeners();
-		ParserErrorListener errorListener = new ParserErrorListener();
-		parser.addErrorListener(errorListener);
-		ParserRuleContext tree = parser.template();
-		return new ParsingResult(parser, tree, errorListener);
-	}
+    @NotNull
+    @Override
+    public TextAttributesKey[] getTokenHighlights(IElementType tokenType) {
+        TextAttributesKey key;
 
-	@Override
-	public void highlightTree(ParserRuleContext tree, Parser parser) {
-		final Collection<ParseTree> ids = XPath.findAll(tree, "//primary/ID", parser);
-		for (ParseTree id : ids) {
-			TerminalNode tnode = (TerminalNode)id;
-			highlightToken(tnode.getSymbol(), new TextAttributesKey[]{ST_ID});
-		}
-	}
+        if (KEYWORDS.contains(tokenType)) {
+            key = DefaultLanguageHighlighterColors.KEYWORD;
+        } else if (STTokenTypes.getTokenElementType(STLexer.ID).equals(tokenType)) {
+            key = ST_ID;
+        } else if (STTokenTypes.getTokenElementType(STLexer.STRING).equals(tokenType)
+                || STTokenTypes.getTokenElementType(STLexer.TEXT).equals(tokenType)) {
+            key = STGroup_TEMPLATE_TEXT;
+        } else if (STTokenTypes.getTokenElementType(STLexer.TMPL_COMMENT).equals(tokenType)) {
+            key = DefaultLanguageHighlighterColors.BLOCK_COMMENT;
+//        } else if (STTokenTypes.getTokenElementType(STLexer.SUB_ERR_CHAR).equals(tokenType)) {
+//            key = HighlighterColors.BAD_CHARACTER;
+        } else {
+            return TextAttributesKey.EMPTY_ARRAY;
+        }
 
-	@NotNull
-	@Override
-	public TextAttributesKey[] getAttributesKey(Token t) {
-		int tokenType = t.getType();
-		TextAttributesKey key;
-		switch ( tokenType ) {
-			case STLexer.IF:
-			case STLexer.ELSE:
-//			case STLexer.REGION_END:
-			case STLexer.TRUE:
-			case STLexer.FALSE:
-			case STLexer.ELSEIF:
-			case STLexer.ENDIF:
-			case STLexer.SUPER:
-				key = DefaultLanguageHighlighterColors.KEYWORD;
-				break;
-			case STLexer.ID:
-				key = ST_ID;
-				break;
-			case STLexer.STRING:
-			case STLexer.TEXT:
-				key = STGroup_TEMPLATE_TEXT;
-				break;
-//			case STLexer.COMMENT:
-//				key = DefaultLanguageHighlighterColors.LINE_COMMENT;
-//				break;
-//			case STLexer.ERROR_TYPE :
-//				key = HighlighterColors.BAD_CHARACTER;
-//				break;
-			default:
-				return EMPTY;
-		}
-		return new TextAttributesKey[]{key};
-	}
+        return new TextAttributesKey[]{key};
+    }
 }
