--- src/main/java/org/antlr/jetbrains/st4plugin/parsing/ParsingUtils.java
@@ -1,14 +1,0 @@
-package org.antlr.jetbrains.st4plugin.parsing;
-
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.CommonTokenStream;
-
-public class ParsingUtils {
-	public static CommonTokenStream tokenize(String text) {
-		ANTLRInputStream input = new ANTLRInputStream(text);
-		STGLexer lexer = new STGLexer(input);
-		CommonTokenStream tokens = new CommonTokenStream(lexer);
-		tokens.fill();
-		return tokens;
-	}
-}
