+++ src/main/resources/META-INF/plugin.xml
@@ -50,10 +50,9 @@
       <structureViewBuilder factoryClass="org.antlr.jetbrains.st4plugin.structview.STGroupStructureViewBuilderProvider"/>
       <additionalTextAttributes scheme="Default" file="colorSchemes/STGroupDefault.xml"/>
    	  <additionalTextAttributes scheme="Darcula" file="colorSchemes/STGroupDarcula.xml"/>
-<!--      <highlightingPassFactory implementation="org.antlr.jetbrains.st4plugin.highlight.STGroupHighlightingPassFactory"/>-->
 
-      <lang.syntaxHighlighterFactory language="STGroup" implementationClass="org.antlr.jetbrains.st4plugin.highlight.NewSTGroupSyntaxHighlighterFactory"/>
-      <lang.syntaxHighlighterFactory language="ST" implementationClass="org.antlr.jetbrains.st4plugin.highlight.NewSTSyntaxHighlighterFactory"/>
+      <lang.syntaxHighlighterFactory language="STGroup" implementationClass="org.antlr.jetbrains.st4plugin.highlight.STGroupSyntaxHighlighterFactory"/>
+      <lang.syntaxHighlighterFactory language="ST" implementationClass="org.antlr.jetbrains.st4plugin.highlight.STSyntaxHighlighterFactory"/>
       <annotator language="STGroup" implementationClass="org.antlr.jetbrains.st4plugin.highlight.STGroupSemanticHighlightAnnotator"/>
       <lang.parserDefinition language="STGroup" implementationClass="org.antlr.jetbrains.st4plugin.psi.STGroupParserDefinition"/>
       <lang.parserDefinition language="ST" implementationClass="org.antlr.jetbrains.st4plugin.psi.STParserDefinition"/>
